moved getAllEvents to lib folder
